add transition composition and refactor
