+++ src/card.css
@@ -1,0 +1,62 @@
+.card-info {
+  z-index: 999;
+  margin: 0.75em;
+
+  padding: 2.5em;
+  border: 1px solid #ccc;
+  border-radius: 50px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+  background-color: aliceblue;
+
+  overflow-y: scroll;
+  /* height: 90vh; */
+  max-height: 90vh;
+  min-height: 20vh;
+}
+
+.card-title {
+  text-shadow: -1px -1px 2px rgba(0, 0, 0, 0.2), 2px 2px 3px rgb(202, 202, 202);
+}
+
+/* scrollbar */
+::-webkit-scrollbar {
+  width: 15px;
+}
+
+/* Track */
+::-webkit-scrollbar-track {
+  border-radius: 8px;
+  background-color: #e7e7e7;
+  border: 1px solid #cacaca;
+
+  display: none;
+}
+
+/* Handle */
+::-webkit-scrollbar-thumb {
+  border-radius: 8px;
+  border: 3px solid transparent;
+  background-clip: content-box;
+  background-color: #d55959;
+}
+
+/* to fake the height of the thumb, I used margins for the top and bottom 'end pieces' of the track */
+::-webkit-scrollbar-track-piece:end {
+  background: transparent;
+  margin-bottom: 45px;
+}
+
+::-webkit-scrollbar-track-piece:start {
+  background: transparent;
+  margin-top: 45px;
+}
+
+@media (max-width: 700px) {
+  .cols {
+    flex-flow: column;
+  }
+
+  .intro {
+    margin: 1em;
+  }
+}
