+++ src/routes/cards/intro/+page.svelte
@@ -1,0 +1,28 @@
+<script>
+  import TransitionCardWrapper from "../../../components/TransitionCardWrapper.svelte";
+</script>
+
+<TransitionCardWrapper>
+  <article class="card-info">
+    <h1 class="card-title">Hello / こんにちわ</h1>
+    <h2>
+      Come join me in climbing <a
+        href="https://en.wikipedia.org/wiki/Mount_Fuji">Mt. Fuji</a
+      > with this interactive experience.
+    </h2>
+    <h2 class="subtitle">Use the mouse to navigate the world.</h2>
+    <h3>Click + Drag to rotate</h3>
+    <h3>Right Click + Drag to pan</h3>
+    <h2 class="subtitle">
+      Click on the flag to get started and follow the path 🔴.
+    </h2>
+  </article>
+</TransitionCardWrapper>
+
+<style>
+  @import "../../../card.css";
+
+  .card-info {
+    max-width: 500px;
+  }
+</style>
