+++ static/assets/MtFuji.glb
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b452672aa290b68a4b20ecfc4e59a57943240690bc8b9405a88ee8dc8843f18e
-size 3095144
+oid sha256:007a17f98e3f530bc3bcb9d0c1dcae84b173e4d1638a9ff33f490ff663ba03c5
+size 3094100
